Fix an x86-64-only empty-profile bug and the stale installer version
Two defects, both found by an audit rather than by anything failing.

dsp/VoiceAnalyzer.h -- UBSan flagged the cast in analyze():

    n = std::min (n, (int) (fs * maxSec));

maxSec is caller-supplied, and the "no cap" callers pass 1e9
(test/profile_dump.cpp at four sites, test/offline_test.cpp at one), so the
product is ~4.8e13 and the conversion to int is undefined. The two
architectures this ships on disagree about what it does: arm64 saturates to
INT_MAX, which happens to be the intended "no cap", while x86-64 yields
INT_MIN -- n goes negative, the length guard below rejects it, and analyze()
returns an empty profile. Verified by compiling the same expression both
ways: arm64 gives 480000, x86-64 gives -2147483648.

Nothing in the shipped plugin passes a maxSec, so the AU/VST3/Standalone
audio path was never affected. What was affected is the built-in target
catalog generator and one offline_test case, on exactly the platforms CI
builds -- the Universal binary's x86-64 slice and Windows. The test case
guards with `if (! p.valid()) continue;`, so it would have gone on passing
while silently measuring nothing.

Doing the min in double means the value being cast is always <= n.

.github/workflows/build.yml -- the installer's --version had been bumped by
hand alongside the project version since v0.63.x and was missed for v0.67.0,
so the package claimed 0.66.0. macOS compares this when deciding whether a
package is an upgrade, so it is not cosmetic. Read from CMakeLists instead,
with a `test -n` so a future rename fails the build loudly rather than
quietly passing pkgbuild an empty version.

Verified: offline_test / protection_test / spatial_test all pass, offline_test
under ASan+UBSan now reports zero findings (was one), Release build of AU,
VST3 and Standalone clean, VoxMorphUiShot ALL PASS. The engine is untouched.

+++ dsp/VoiceAnalyzer.h
@@ -171,7 +171,17 @@ public:
     {
         VoiceProfile out;
         constexpr int W = 2048, hop = 1024, N = 4096, NB = N / 2;
-        n = std::min (n, (int) (fs * maxSec));
+        // The comparison has to happen in double, not after the cast.
+        // maxSec is caller-supplied and the "no cap" callers pass 1e9
+        // (test/profile_dump.cpp, and one case in offline_test), so
+        // fs * maxSec is ~4.8e13 -- far outside int. Converting that to int
+        // is undefined behaviour, and the two architectures this ships on
+        // disagree about it: arm64 saturates to INT_MAX (which happens to be
+        // the intended "no cap"), while x86-64 yields INT_MIN, so n went
+        // NEGATIVE and analyze() returned an empty profile on exactly the
+        // platforms the CI builds for. Doing the min first means the value
+        // being cast is always <= n and therefore always in range.
+        n = (int) std::min ((double) n, fs * maxSec);
         const int maxLag = (int) (fs / 60.0), minLag = std::max (2, (int) (fs / 500.0));
         if (n < W + maxLag + 8) return out;
